feat: migrate from SDL2 to SDL3 across all platforms
Complete SDL2 to SDL3 migration covering the entire codebase:

Core migration (115 files, ~4200 insertions, ~3400 deletions):
- Replace all SDL2 APIs with SDL3 equivalents (types, functions, constants)
- Handle SDL3 bool return type changes (was int 0=success, now bool true=success)
- Migrate SDL3 Audio (SDL_AudioStream API), Gamepad, Window, and Renderer APIs
- Fix SDL_WindowFlags to Uint64 (was Uint32, caused silent truncation)
- Add NULL-deref protection for SDL_GetPixelFormatDetails() (8 files)
- Hoist SDL_GetPixelFormatDetails() out of pixel loops (4 performance sites)
- Add SDL3 audio error checks for AudioStream API (3 files)

Build system:
- CMake: Replace SDL2/SDL2_image targets with SDL3/SDL3_image
- Remove USE_IMGUI option (ImGui is now the only GUI, always enabled)
- Android: FetchContent pins SDL 3.2.8, SDL3_image 3.2.4
- vcpkg.json: sdl2/sdl2-image -> sdl3/sdl3-image
- Brewfile: sdl2/sdl2_image -> sdl3/sdl3_image
- Remove cmake/FindSDL2.cmake and cmake/FindSDL2_image.cmake

Android:
- Replace all SDL2 Java files with SDL3 versions (11 files)
- Add SDLDummyEdit.java and SDLInputConnection.java (new in SDL3)
- Update AmiberryActivity.java for SDL3 (getLibraries override, SDL_main.h)

Libretro:
- Rewrite SDL3 compatibility shim (sdl_compat.h: 1079 lines, sdl_stub.cpp: 1393 lines)
- Add SDL3 redirect headers (SDL3/SDL.h, SDL_main.h, SDL_mutex.h, SDL_opengl.h, SDL_opengles2.h)
- Update Makefile: gnu++17, gnu99, arm64 detection, macOS linker flags
- Add macOS libretro build target to CI workflow

Packaging:
- DEB/RPM package dependencies updated to SDL3
- Flatpak: runtime 24.08 -> 25.08, add SDL3 3.4.2 and SDL3_image 3.4.0 modules
- Flatpak: update libserialport 0.1.1 -> 0.1.2, portmidi v2.0.4 -> v2.0.8

Code cleanup:
- Rename SDL2-prefixed functions to amiberry_* or sdl_* equivalents
- Remove orphaned SDL2 function declarations
- Update ~30 stale SDL2 references in comments/logs
- Remove dead code (traps.cpp -2 comparison, gfx_colors.cpp duplicate line)
- Add controller map entries for MISC2-MISC6

Fixes #1618

+++ src/osdep/writelog.cpp
@@ -591,7 +591,7 @@ TCHAR* write_log_get_ts(void)
 	if (!vsync_counter)
 		return nullptr;
 
-	Uint32 ticks = SDL_GetTicks();
+	uint64_t ticks = SDL_GetTicks();
 	time_t seconds = ticks / 1000;
 	int milliseconds = ticks % 1000;
